requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==3.2
django-channels==0.7.0
django-countries==7.5.1
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.14.0
# djangorestframework-filters
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.5.0
drf-spectacular==0.26.2
fcm-django==2.0.0
filetype==1.2.0
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
inflection==0.5.1
jsonschema==4.17.3
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
rave-python==1.4.0
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
scikit-learn==1.3.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stripe==6.0.0
tomli==2.0.1
twilio==8.4.0
Twisted==23.8.0
# twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
